Move update state function into module
