+++ chromium/manifest.json
@@ -13,7 +13,8 @@
 						"background-scripts/validation.js",
 						"background-scripts/modules/ip_utils.js",
 						"background-scripts/modules/browser_session.js",
-						"background-scripts/modules/ssl_codes.js",
+            "background-scripts/modules/ssl_codes.js",
+            "background-scripts/modules/state.js",
             "wasm/https_everywhere_lib_wasm.js",
             "background-scripts/wasm.js",
             "background-scripts/update_channels.js",
